Update filltemplate to fillTemplate
