--- packages/aws/stepDefinitions/modules.js
@@ -1,2 +1,0 @@
-require('@ln-maf/validations')
-require('@ln-maf/core/parameter_types')
